Change time.Duration(int64(0)) -> time.Duration(0)

+++ configmap/tools.go
@@ -48,7 +48,7 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case string:
 		cm, err := h.Get(val)
 		if err != nil {
-			return time.Duration(int64(0)), err
+			return time.Duration(0), err
 		}
 		return time.Now().Sub(cm.CreationTimestamp.Time), nil
 	case *corev1.ConfigMap:
@@ -56,6 +56,6 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case corev1.ConfigMap:
 		return time.Now().Sub(val.CreationTimestamp.Time), nil
 	default:
-		return time.Duration(int64(0)), ErrInvalidToolsType
+		return time.Duration(0), ErrInvalidToolsType
 	}
 }
