+++ persistentvolumeclaim/tools.go
@@ -140,7 +140,7 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case string:
 		pvc, err := h.Get(val)
 		if err != nil {
-			return time.Duration(int64(0)), err
+			return time.Duration(0), err
 		}
 		return time.Now().Sub(pvc.CreationTimestamp.Time), nil
 	case *corev1.PersistentVolumeClaim:
@@ -148,7 +148,7 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case corev1.PersistentVolumeClaim:
 		return time.Now().Sub(val.CreationTimestamp.Time), nil
 	default:
-		return time.Duration(int64(0)), ErrInvalidToolsType
+		return time.Duration(0), ErrInvalidToolsType
 	}
 }
